update paperpal.py
cleaned up the main class and removed the temp file
